# api/swagger.json
{"info":{"title":"Hello World","version":"1.0.0","description":"A sample API"},"host":"localhost:3000","basePath":"/","swagger":"2.0","paths":{"/version/{versionId}/figure":{"get":{"description":"Get all Figures for a Version","tags":["Figures"],"parameters":[{"$ref":"#/parameters/versionId"}],"produces":["application/json"],"responses":{"200":{"description":"Returns all Figures for a Version","schema":{"type":"array","items":{"$ref":"#/definitions/Scene"}}}}},"post":{"description":"Create a Figure","tags":["Figures"],"parameters":[{"$ref":"#/parameters/versionId"},{"$ref":"#/parameters/figureName"},{"$ref":"#/parameters/geometry"},{"$ref":"#/parameters/figureProperties"},{"$ref":"#/parameters/figureType"}],"consumes":["application/x-www-form-urlencoded"],"produces":["application/json"],"responses":{"200":{"description":"Returns ID of new Figure","schema":{"type":"object"}}}}},"/figure/{figureId}":{"get":{"description":"Get Figure by ID","tags":["Figures"],"parameters":[{"$ref":"#/parameters/figureId"}],"produces":["application/json"],"responses":{"200":{"description":"Returns Figure object","schema":{"type":"array","items":{"$ref":"#/definitions/Figure"}}}}},"put":{"description":"Edit a Figure","tags":["Figures"],"parameters":[{"$ref":"#/parameters/figureId"},{"$ref":"#/parameters/figureName"},{"$ref":"#/parameters/geometry"},{"$ref":"#/parameters/figureProperties"},{"$ref":"#/parameters/figureType"}],"consumes":["application/x-www-form-urlencoded"],"produces":["application/json"],"responses":{"200":{"description":"Returns ID of edited Version","schema":{"type":"object"}}}},"delete":{"description":"Delete a Figure","tags":["Figures"],"parameters":[{"$ref":"#/parameters/figureId"}],"produces":["application/json"],"responses":{"200":{"description":"Returns ID of deleted Figure","schema":{"type":"object"}}}},"patch":{"description":"Reorder children of a Figure Folder","tags":["Figures"],"parameters":[{"$ref":"#/parameters/figureId"},{"$ref":"#/parameters/patchOperations"}],"consumes":["application/x-www-form-urlencoded"],"produces":["application/json"],"responses":{"200":{"description":"Returns ids of Figure children in the new order","schema":{"type":"object"}}}}},"/figure/{figureId}/children":{"get":{"description":"Get children of Figure","tags":["Figures"],"parameters":[{"$ref":"#/parameters/figureId"}],"produces":["application/json"],"responses":{"200":{"description":"Returns all children of Figure","schema":{"type":"array","items":{"$ref":"#/definitions/Figure"}}}}}},"/group":{"get":{"description":"Get All Groups","tags":["Groups"],"produces":["application/json"],"responses":{"200":{"description":"Returns all Groups","schema":{"type":"array","items":{"$ref":"#/definitions/Group"}}}}},"post":{"description":"Create a Group","tags":["Groups"],"parameters":[{"$ref":"#/parameters/groupName"},{"$ref":"#/parameters/groupDescription"},{"$ref":"#/parameters/city"},{"$ref":"#/parameters/country"}],"consumes":["application/x-www-form-urlencoded"],"produces":["application/json"],"responses":{"200":{"description":"Returns ID of created Group","schema":{"type":"object"}}}}},"/group/{groupId}":{"get":{"description":"Get a Group by ID","tags":["Groups"],"parameters":[{"$ref":"#/parameters/groupId"}],"produces":["application/json"],"responses":{"200":{"description":"Returns a Group","schema":{"type":"array","items":{"$ref":"#/definitions/Group"}}}}},"put":{"description":"Edit a Group","tags":["Groups"],"parameters":[{"$ref":"#/parameters/groupId"},{"$ref":"#/parameters/groupName"},{"$ref":"#/parameters/groupDescription"},{"$ref":"#/parameters/city"},{"$ref":"#/parameters/country"}],"consumes":["application/x-www-form-urlencoded"],"produces":["application/json"],"responses":{"200":{"description":"Returns ID of new Group","schema":{"type":"object"}}}}},"/group/{groupId}/scene":{"get":{"description":"Get all Scenes for a Group","tags":["Scenes"],"parameters":[{"$ref":"#/parameters/groupId"}],"produces":["application/json"],"responses":{"200":{"description":"Returns all Scenes for a Group","schema":{"type":"array","items":{"$ref":"#/definitions/Scene"}}}}},"post":{"description":"Create a Scene","tags":["Scenes"],"parameters":[{"$ref":"#/parameters/groupId"},{"$ref":"#/parameters/sceneName"},{"$ref":"#/parameters/sceneDescription"}],"consumes":["application/x-www-form-urlencoded"],"produces":["application/json"],"responses":{"200":{"description":"Returns ID of new Scene","schema":{"type":"object"}}}}},"/scene/{sceneId}":{"get":{"description":"Get Scene by ID","tags":["Scenes"],"parameters":[{"$ref":"#/parameters/sceneId"}],"produces":["application/json"],"responses":{"200":{"description":"Returns a Scene","schema":{"type":"array","items":{"$ref":"#/definitions/Scene"}}}}},"put":{"description":"Edit a Scene","tags":["Scenes"],"parameters":[{"$ref":"#/parameters/sceneId"},{"$ref":"#/parameters/sceneName"},{"$ref":"#/parameters/sceneDescription"}],"consumes":["application/x-www-form-urlencoded"],"produces":["application/json"],"responses":{"200":{"description":"Returns ID of edited Scene","schema":{"type":"object"}}}}},"/scene/{sceneId}/version":{"get":{"description":"Get all Versions for a Scene","tags":["Versions"],"parameters":[{"$ref":"#/parameters/sceneId"}],"produces":["application/json"],"responses":{"200":{"description":"Returns all Versions for a Scene","schema":{"type":"array","items":{"$ref":"#/definitions/Scene"}}}}},"post":{"description":"Create a Scene","tags":["Versions"],"parameters":[{"$ref":"#/parameters/sceneId"},{"$ref":"#/parameters/versionName"},{"$ref":"#/parameters/figures"},{"$ref":"#/parameters/layers"}],"consumes":["application/x-www-form-urlencoded"],"produces":["application/json"],"responses":{"200":{"description":"Returns ID of new Version","schema":{"type":"object"}}}}},"/version/{sceneId}":{"get":{"description":"Get Version by ID","tags":["Versions"],"parameters":[{"$ref":"#/parameters/versionId"}],"produces":["application/json"],"responses":{"200":{"description":"Returns Version object","schema":{"type":"array","items":{"$ref":"#/definitions/Version"}}}}},"put":{"description":"Edit a Version","tags":["Versions"],"parameters":[{"$ref":"#/parameters/versionId"},{"$ref":"#/parameters/versionName"},{"$ref":"#/parameters/layers"}],"consumes":["application/x-www-form-urlencoded"],"produces":["application/json"],"responses":{"200":{"description":"Returns ID of edited Version","schema":{"type":"object"}}}}}},"definitions":{"Figure":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","maxLength":255},"type":{"type":"integer","format":"int32"},"geometry":{"type":"string","maxLength":255},"properties":{"type":"string","maxLength":255},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"version":{"type":"integer","format":"int32"},"parentId":{"type":"integer","format":"int32"},"previousId":{"type":"integer","format":"int32"},"versionId":{"type":"integer","format":"int32"}},"required":["id","createdAt","updatedAt","version","versionId"]},"Group":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","maxLength":255},"city":{"type":"string","maxLength":255},"country":{"type":"string","maxLength":255},"description":{"type":"string","maxLength":255},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}},"required":["id","createdAt","updatedAt"]},"Scene":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","maxLength":255},"description":{"type":"string","maxLength":255},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"groupId":{"type":"integer","format":"int32"},"defaultVersionId":{"type":"integer","format":"int32"}},"required":["id","createdAt","updatedAt","groupId"]},"Version":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string","maxLength":255},"layers":{"type":"string","maxLength":255},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"sceneId":{"type":"integer","format":"int32"}},"required":["id","createdAt","updatedAt","sceneId"]}},"responses":{},"parameters":{"figureId":{"name":"figureId","type":"integer","in":"path","required":true,"description":"Numeric ID of the figure"},"figureName":{"name":"name","description":"Name of figure","type":"string","in":"formData"},"figureDescription":{"name":"description","description":"Description of figure","type":"string","in":"formData"},"figureType":{"name":"type","description":"Numeric ID of the figures type","type":"integer","in":"formData"},"figurePrevious":{"name":"previousId","description":"Numeric ID of the Figure behing which to insert this figure","type":"string","in":"formData"},"figureParent":{"name":"parentId","description":"Numeric ID of the Folder in which to insert this figure","type":"string","in":"formData"},"geometry":{"name":"geometry","description":"Stringified geoJSON representing the figure geometry","type":"string","in":"formData"},"figureProperties":{"name":"properties","description":"Stringified properties object as defined by Contour","type":"string","in":"formData"},"patchOperations":{"name":"operations","description":"Array of JSON patch operations to apply to Figure.children. http://jsonpatch.com/   \n e.g. [{\"op\":\"replace\",\"path\":\"/12\",\"value\":3},{\"op\":\"replace\",\"path\":\"/7\",\"value\":null},{\"op\":\"replace\",\"path\":\"/3\",\"value\":6}]","type":"string","in":"formData"},"groupId":{"name":"groupId","type":"integer","in":"path","required":true,"description":"Numeric ID of the group"},"groupName":{"name":"name","description":"Name of group","type":"string","in":"formData"},"groupDescription":{"name":"description","description":"Description of group","type":"string","in":"formData"},"city":{"name":"city","description":"City in which group is based","type":"string","in":"formData"},"country":{"name":"country","description":"Country in which group is based","type":"string","in":"formData"},"sceneId":{"name":"sceneId","type":"integer","in":"path","required":true,"description":"Numeric ID of the scene"},"sceneName":{"name":"name","description":"Name of scene","type":"string","in":"formData"},"sceneDescription":{"name":"description","description":"Description of scene","type":"string","in":"formData"},"versionId":{"name":"versionId","type":"integer","in":"path","required":true,"description":"Numeric ID of the version"},"versionName":{"name":"name","description":"Name of version","type":"string","in":"formData"},"versionDescription":{"name":"description","description":"Description of version","type":"string","in":"formData"},"layers":{"name":"layers","description":"Comma separated list of layer IDs. e.g. 'Google_Hybrid,HnP_Hazards,Recent_Earthquakes'","type":"string","in":"formData"},"figures":{"name":"figures","description":"Stringified JSON representing the full figure treee","type":"string","in":"formData"}},"securityDefinitions":{},"tags":[]}